requirements.txt
Django==2.1.7
psycopg2==2.8.2
djangorestframework==3.9.2

